Account aggregate handles MoneyDeposit command

+++ examples/bank/pkg/event/money_withdrawn.go
@@ -4,8 +4,8 @@ import "github.com/screwyprof/cqrs/examples/bank/pkg/domain"
 
 type MoneyWithdrawn struct {
 	ID      domain.Identifier
-	Amount  int
-	Balance int
+	Amount  int64
+	Balance int64
 }
 
 func (e MoneyWithdrawn) EventType() string {
